refactor: Consolidate unit tests under /app/tests
This commit consolidates all Vitest-based unit and component tests into a single `/app/tests` directory.

Changes:
- Moved all `*.test.ts` and `*.test.tsx` files from `/app/src/**` to `/app/tests/`.
- Created a new directory structure: `/app/tests/unit`, `/app/tests/components`, and `/app/tests/setup`.
- Updated `app/vite.config.ts` to discover tests in the new location.
- Updated `app/tsconfig.json` to include the new test directory for type checking.
- Updated `AGENTS.md` to reflect the new test location.
- Fixed all relative import paths in the moved test files.

Verification:
- All quality gates passed successfully.
